Only compile sha2 dep on x86
