+++ Cargo.toml
@@ -13,10 +13,10 @@ categories = ["cryptography::cryptocurrencies"]
 [dependencies]
 lazy_static = { version = "1.1", optional = true }
 ring = "0.17"
-sha2 = "0.10"
 
 [target.'cfg(target_arch = "x86_64")'.dependencies]
 cpufeatures = "0.2"
+sha2 = "0.10"
 
 [dev-dependencies]
 rustc-hex = "2"
